[Daikin] Rename OpenDaikin to Daikin
Note that this will break any current users of OpenDaikin as the item namespaces have changed.
