+++ addons/binding/org.openhab.binding.daikin/.classpath
