+++ addons/binding/org.openhab.binding.daikin/ESH-INF/binding/binding.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<binding:binding id="daikin"
+				xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+				xmlns:binding="http://eclipse.org/smarthome/schemas/binding/v1.0.0"
+				xsi:schemaLocation="http://eclipse.org/smarthome/schemas/binding/v1.0.0 http://eclipse.org/smarthome/schemas/binding-1.0.0.xsd">
+
+	<name>Daikin Binding</name>
+	<description>This is the binding for Daikin A/C units.</description>
+	<author>Tim Waterhouse</author>
+
+</binding:binding>
