+++ addons/binding/org.openhab.binding.daikin/NOTICE
