+++ addons/binding/org.openhab.binding.daikin/OSGI-INF/.gitignore
