+++ addons/binding/org.openhab.binding.daikin/build.properties
