--- addons/binding/org.openhab.binding.opendaikin/ESH-INF/binding/binding.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<binding:binding id="opendaikin"
-				xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-				xmlns:binding="http://eclipse.org/smarthome/schemas/binding/v1.0.0"
-				xsi:schemaLocation="http://eclipse.org/smarthome/schemas/binding/v1.0.0 http://eclipse.org/smarthome/schemas/binding-1.0.0.xsd">
-
-	<name>OpenDaikin Binding</name>
-	<description>This is the binding for OpenDaikin.</description>
-	<author>Tim Waterhouse</author>
-
-</binding:binding>
