--- addons/binding/org.openhab.binding.opendaikin/ESH-INF/thing/thing-types.xml
@@ -1,118 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<thing:thing-descriptions bindingId="opendaikin" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns:thing="http://eclipse.org/smarthome/schemas/thing-description/v1.0.0"
-	xsi:schemaLocation="http://eclipse.org/smarthome/schemas/thing-description/v1.0.0 http://eclipse.org/smarthome/schemas/thing-description-1.0.0.xsd">
-
-	<thing-type id="ac_unit">
-		<label>Daikin AC Unit</label>
-		<description>Daikin AC Unit</description>
-
-		<channels>
-			<channel id="power" typeId="acunit-power" />
-			<channel id="settemp" typeId="acunit-settemp" />
-			<channel id="indoortemp" typeId="acunit-indoortemp" />
-			<channel id="outdoortemp" typeId="acunit-outdoortemp" />
-			<channel id="humidity" typeId="acunit-humidity" />
-			<channel id="mode" typeId="acunit-mode" />
-			<channel id="fanspeed" typeId="acunit-fan" />
-			<channel id="fandir" typeId="acunit-fandir" />
-		</channels>
-
-		<config-description>
-			<parameter name="host" type="text" required="true">
-				<label>Host</label>
-				<description>The Host address of the Daikin AC unit.</description>
-				<context>network_address</context>
-			</parameter>
-			<parameter name="refresh" type="integer" required="false">
-				<label>Refresh Interval</label>
-				<description>The number of seconds between fetches of the AC unit state.</description>
-				<default>60</default>
-			</parameter>
-		</config-description>
-	</thing-type>
-
-	<channel-type id="acunit-power">
-		<item-type>Switch</item-type>
-		<label>Power</label>
-		<description>Power for the AC unit</description>
-	</channel-type>
-
-	<channel-type id="acunit-settemp">
-		<item-type>Number:Temperature</item-type>
-		<label>Set Point</label>
-		<description>The set point temperature</description>
-		<category>Temperature</category>
-		<state pattern="%.1f %unit%" step="0.5" />
-	</channel-type>
-
-	<channel-type id="acunit-indoortemp">
-		<item-type>Number:Temperature</item-type>
-		<label>Indoor Temperature</label>
-		<description>The indoor temperature</description>
-		<category>Temperature</category>
-		<state readOnly="true" pattern="%.1f %unit%" step="0.5" />
-	</channel-type>
-
-	<channel-type id="acunit-outdoortemp">
-		<item-type>Number:Temperature</item-type>
-		<label>Outdoor Temperature</label>
-		<description>The outdoor temperature</description>
-		<category>Temperature</category>
-		<state readOnly="true" pattern="%.1f %unit%" step="0.5" />
-	</channel-type>
-
-	<channel-type id="acunit-humidity" advanced="true">
-		<item-type>Number</item-type>
-		<label>Indoor Humidity</label>
-		<description>The indoor humidity as measured by the A/C unit</description>
-		<state readOnly="true" pattern="%.0f%%" />
-	</channel-type>
-
-	<channel-type id="acunit-mode">
-		<item-type>String</item-type>
-		<label>Mode</label>
-		<description>Current mode of the AC unit</description>
-		<state>
-			<options>
-				<option value="AUTO">auto</option>
-				<option value="DEHUMIDIFIER">dehumidifier</option>
-				<option value="COLD">cooling</option>
-				<option value="HEAT">heating</option>
-				<option value="FAN">fan</option>
-			</options>
-		</state>
-	</channel-type>
-
-	<channel-type id="acunit-fan" advanced="true">
-		<item-type>String</item-type>
-		<label>Fan</label>
-		<description>Current fan speed setting of the AC unit</description>
-		<state>
-			<options>
-				<option value="AUTO">auto</option>
-				<option value="SILENCE">silence</option>
-				<option value="LEVEL_1">level 1</option>
-				<option value="LEVEL_2">level 2</option>
-				<option value="LEVEL_3">level 3</option>
-				<option value="LEVEL_4">level 4</option>
-				<option value="LEVEL_5">level 5</option>
-			</options>
-		</state>
-	</channel-type>
-	
-	<channel-type id="acunit-fandir" advanced="true">
-		<item-type>String</item-type>
-		<label>Fan Swing</label>
-		<description>Current fan swing setting of the AC unit</description>
-		<state>
-			<options>
-				<option value="STOPPED">stopped</option>
-				<option value="VERTICAL">vertical</option>
-				<option value="HORIZONTAL">horizontal</option>
-				<option value="VERTICAL_AND_HORIZONTAL">vertical and horizontal</option>
-			</options>
-		</state>
-	</channel-type>
-
-</thing:thing-descriptions>
