--- addons/binding/org.openhab.binding.opendaikin/pom.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
-		xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>org.openhab.binding</groupId>
-		<artifactId>pom</artifactId>
-		<version>2.5.0-SNAPSHOT</version>
-	</parent>
-
-	<artifactId>org.openhab.binding.opendaikin</artifactId>
-
-	<name>OpenDaikin Binding</name>
-	<packaging>eclipse-plugin</packaging>
-
-</project>
